Reorder Dockerfile
Attempt to put Dockerfile commands in order of least likely to change.

As each line of the Dockerfile becomes a layer of the docker image,
we should put the lines most likely to change last, so when they do change
only the last layer needs to be updated. E.g. if you change a script in the
/bin folder.